pyproject.toml

[tool.poetry]
name = "python-qweather"
version = "0.1.3"
description = "Python API wrapper for https://qweather.com"
authors = ["dofine <dofine@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
keywords = ["weather"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
repository = "https://github.com/dofine/python-qweather"


[tool.poetry.dependencies]
python = "^3.7"
pytest-asyncio = "^0.18.3"

[tool.poetry.dev-dependencies]



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.bumpver]
current_version = "0.1.3"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'version = "{version}"',
]
"python_qweather/__init__.py" = [
    '__version__ = "{version}"'
]

